+++ packages/lib/test/actionMiddlewares/undoMiddleware.test.ts
@@ -13,6 +13,7 @@ import {
   UndoManager,
   undoMiddleware,
   UndoStore,
+  withoutUndo,
   _async,
   _await,
 } from "../../src"
@@ -60,6 +61,23 @@ class R extends Model({
   p: prop(() => new P({})),
 }) {}
 
+function expectUndoManagerRedoToBe(manager: UndoManager, undoLevels: number, redoLevels: number) {
+  expect(manager.canUndo).toBe(undoLevels > 0)
+  expect(manager.undoLevels).toBe(undoLevels)
+  expect(manager.undoQueue.length).toBe(undoLevels)
+
+  expect(manager.canRedo).toBe(redoLevels > 0)
+  expect(manager.redoLevels).toBe(redoLevels)
+  expect(manager.redoQueue.length).toBe(redoLevels)
+
+  if (undoLevels <= 0) {
+    expect(() => manager.undo()).toThrow("nothing to undo")
+  }
+  if (redoLevels <= 0) {
+    expect(() => manager.redo()).toThrow("nothing to redo")
+  }
+}
+
 test("undoMiddleware - sync", () => {
   const r = new R({})
   const p = r.p
@@ -76,20 +94,7 @@ test("undoMiddleware - sync", () => {
   }
 
   function expectUndoRedoToBe(undoLevels: number, redoLevels: number) {
-    expect(manager.canUndo).toBe(undoLevels > 0)
-    expect(manager.undoLevels).toBe(undoLevels)
-    expect(manager.undoQueue.length).toBe(undoLevels)
-
-    expect(manager.canRedo).toBe(redoLevels > 0)
-    expect(manager.redoLevels).toBe(redoLevels)
-    expect(manager.redoQueue.length).toBe(redoLevels)
-
-    if (undoLevels <= 0) {
-      expect(() => manager.undo()).toThrow("nothing to undo")
-    }
-    if (redoLevels <= 0) {
-      expect(() => manager.redo()).toThrow("nothing to redo")
-    }
+    expectUndoManagerRedoToBe(manager, undoLevels, redoLevels)
   }
 
   expectUndoRedoToBe(0, 0)
@@ -875,4 +880,31 @@ test("undo-aware substore called from non undo-aware root store", () => {
     ]
   `)
   manager.clearUndo()
+})
+
+test("does not generate steps if using withoutUndo", () => {
+  const r = new R({})
+  const p = r.p
+
+  const manager = undoMiddleware(r, r.undoData)
+  autoDispose(() => manager.dispose())
+
+  function expectUndoRedoToBe(undoLevels: number, redoLevels: number) {
+    expectUndoManagerRedoToBe(manager, undoLevels, redoLevels)
+  }
+
+  expectUndoRedoToBe(0, 0)
+
+  manager.withoutUndo(() => {
+    p.incX(1)
+    p.incX(1)
+  })
+
+  withoutUndo(() => {
+    p.incX(1)
+    p.incX(1)
+  })
+
+  expect(p.x).toBe(4)
+  expectUndoRedoToBe(0, 0)
 })